# design/openapi.yaml
openapi: 3.0.3
info:
  title: rps-xml-tracker API
  version: 0.1.0-draft
  description: >
    API de rastreabilidade de XML (notas fiscais) da RPS Contábil. Entregável de 1ª classe:
    o frontend do rps-maestro (maestro_web) consome esta API numa seção /xml.
    Convenções espelham o maestro: base /api/v1, JWT Bearer (HS256, MAESTRO_JWT_SECRET),
    erros {"error": "..."}, listas {items,total,limit,offset}, datas ISO-8601.
    RASCUNHO para revisão (Fase 1 MVP). Reflete os achados da Fase 0.

servers:
  - url: /api/v1

components:
  securitySchemes:
    bearerAuth:        # leitura (UI do maestro) — papéis admin/operator/viewer
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentHmac:         # ingestão (agente no SRVIMPORT) — HMAC compartilhado, NÃO JWT
      type: apiKey
      in: header
      name: X-Agent-Signature
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
    DocType:
      type: string
      enum: [NFE, NFCE, CTE, NFS, EVENTO, UNKNOWN]
    NotaStatus:
      type: string
      description: >
        Estado derivado da nota. Mapeamento das etapas para as pastas/Athenas.
        Precedência: imported > import_ignored > pending_import > synced > arrived.
        arrived = visto na pasta de chegada (A Sincronizar), aguardando sincronização;
        synced = movido para SINCRONIZADO pelo agent (arquivo posicionado), mas o
          Athenas AINDA NÃO o enxergou;
        pending_import = o poller já viu a chave na TABLISTACHAVEACESSO do Athenas
          com IMPORTADO=0 (Athenas reconheceu, falta importar) — um passo À FRENTE de
          synced, não um sinônimo;
        imported = importado no Athenas (terminal, sucesso);
        import_ignored = IMPORTACAOIGNORADA (terminal, esperado por config da empresa);
        stuck/lost = derivados da fase de alertas (ainda não emitidos — hoje sempre 0).
      enum: [arrived, synced, imported, import_ignored, pending_import, stuck, lost]
    Observation:           # item de ingestão enviado pelo agente
      type: object
      required: [stage, event_type, observed_at]
      properties:
        chave_acesso: { type: string, minLength: 44, maxLength: 44, nullable: true }
        stage:        { type: string, enum: [arrival, sync, import] }
        event_type:   { type: string, example: file_seen }
        observed_at:  { type: string, format: date-time }
        doc_type:     { $ref: '#/components/schemas/DocType' }
        file_path:    { type: string, nullable: true }
        file_hash:    { type: string, nullable: true, description: md5 do conteúdo }
        codigo_empresa: { type: integer, nullable: true }
        codigo_filial:  { type: integer, nullable: true }
        payload:      { type: object, additionalProperties: true, nullable: true }
    Span:
      type: object
      properties:
        stage:       { type: string, enum: [arrival, sync, import] }
        event_type:  { type: string }
        observed_at: { type: string, format: date-time }
        source:      { type: string, example: 'agent:SRVIMPORT' }
        file_path:   { type: string, nullable: true }
    Nota:
      type: object
      properties:
        chave_acesso:     { type: string }
        numero_nota:      { type: string, description: "nNF (número da nota), derivado da chave; ausente p/ NFSe" }
        doc_type:         { $ref: '#/components/schemas/DocType' }
        status:           { $ref: '#/components/schemas/NotaStatus' }
        empresa:          { $ref: '#/components/schemas/EmpresaRef' }
        cnpj_emitente:    { type: string, nullable: true }
        cnpj_destinatario: { type: string, nullable: true }
        arrived_at:       { type: string, format: date-time, nullable: true }
        synced_at:        { type: string, format: date-time, nullable: true }
        pending_at:       { type: string, format: date-time, nullable: true, description: "1º avistamento no Athenas com IMPORTADO=0 (dispara status pending_import)" }
        imported_at:      { type: string, format: date-time, nullable: true }
        import_ignored:   { type: boolean }
        motivo_ignorado:  { type: string, nullable: true }
        valor_total:      { type: number, nullable: true }
        data_emissao:     { type: string, format: date, nullable: true }
        lat_arrival_sync_s: { type: integer, nullable: true }
        lat_sync_import_s:  { type: integer, nullable: true }
    NotaDetail:
      allOf:
        - $ref: '#/components/schemas/Nota'
        - type: object
          properties:
            spans: { type: array, items: { $ref: '#/components/schemas/Span' } }
    EmpresaRef:
      type: object
      properties:
        codigo_empresa: { type: integer }
        codigo_filial:  { type: integer }
        nome:           { type: string, nullable: true }
    # Agregado por empresa (forma FLAT, igual ao que a API retorna).
    EmpresaPendencias:
      type: object
      description: >
        Uma linha por (codigo_empresa, codigo_filial). A linha "Sem empresa" (notas com
        codigo_empresa nulo) vem com codigo_empresa e codigo_filial AUSENTES (omitempty, não
        serializa null) e nome_empresa ausente — detecte por codigo_empresa == null no cliente.
        Para o drill-down dela use GET /notas?sem_empresa=true.
      properties:
        codigo_empresa: { type: integer, description: ausente na linha "Sem empresa" }
        codigo_filial:  { type: integer, description: ausente quando nulo (filial não derivada) ou na linha "Sem empresa" }
        nome_empresa:   { type: string, description: nome da empresa (max() lexical entre as notas); ausente quando vazio }
        in_transit:     { type: integer, description: arrived + synced }
        arrived:        { type: integer }
        synced:         { type: integer }
        imported:       { type: integer }
        import_ignored: { type: integer }
        pending_import: { type: integer }
        stuck:          { type: integer }
        lost:           { type: integer }

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Liveness (sem auth)
      security: []
      responses: { '200': { description: ok } }

  # ───────── Ingestão (agente SRVIMPORT — HMAC, não JWT) ─────────
  /ingest/observations:
    post:
      summary: Recebe um lote de observações do agente (idempotente por dedup_key)
      security: [ { agentHmac: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent, batch]
              properties:
                agent: { type: string, example: 'SRVIMPORT' }
                batch:
                  type: array
                  items: { $ref: '#/components/schemas/Observation' }
      responses:
        '202':
          description: aceito
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: { type: integer }
                  rejected: { type: integer }
        '401': { description: assinatura inválida, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ───────── Leitura (UI do maestro — JWT viewer+) ─────────
  /notas:
    get:
      summary: Lista notas (estado derivado) com filtros
      parameters:
        - { name: status,   in: query, schema: { $ref: '#/components/schemas/NotaStatus' } }
        - { name: doc_type, in: query, schema: { $ref: '#/components/schemas/DocType' } }
        - { name: codigo_empresa, in: query, schema: { type: integer } }
        - { name: codigo_filial,  in: query, schema: { type: integer }, description: "filial exata; combina com codigo_empresa (AND)" }
        - { name: sem_empresa, in: query, schema: { type: boolean }, description: "true = só notas sem empresa identificada (codigo_empresa nulo); drill-down da linha \"Sem empresa\"" }
        - { name: empresa,  in: query, schema: { type: string }, description: nome da empresa (busca parcial) }
        - { name: cnpj,     in: query, schema: { type: string }, description: emitente ou destinatário }
        - { name: q,        in: query, schema: { type: string }, description: chave de acesso (parcial/total) }
        - { name: date_field, in: query, schema: { type: string, enum: [emissao, arrived, synced, imported] }, description: campo de data para o filtro from/to }
        - { name: from,     in: query, schema: { type: string, format: date } }
        - { name: to,       in: query, schema: { type: string, format: date } }
        - { name: limit,    in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset,   in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:  { type: array, items: { $ref: '#/components/schemas/Nota' } }
                  total:  { type: integer }
                  limit:  { type: integer }
                  offset: { type: integer }

  /notas/{chave}:
    get:
      summary: Detalhe de uma nota + timeline de spans (todas as observations da chave)
      parameters:
        - { name: chave, in: path, required: true, schema: { type: string, minLength: 44, maxLength: 44 } }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/NotaDetail' } } } }
        '404': { description: não encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /empresas:
    get:
      summary: Empresas com contagem por status (visão por empresa)
      description: >
        Agregado por empresa/filial: in_transit (arrived+synced), arrived, synced, imported,
        import_ignored, pending_import, stuck, lost, e nome_empresa. Para o drill-down das notas
        de uma empresa, use GET /notas?codigo_empresa=...&status=...
      parameters:
        - { name: pendentes, in: query, schema: { type: boolean }, description: só empresas com itens não-terminais (arrived/synced/pending_import/stuck) }
        - { name: q, in: query, schema: { type: string }, description: "busca por nome da empresa (parcial, case-insensitive)" }
        - { name: sort, in: query, schema: { type: string, enum: [codigo, pendentes] }, description: "ordenação: 'pendentes' = mais pendentes primeiro; default por código" }
        - { name: limit,  in: query, schema: { type: integer, default: 0 }, description: "0 = todas (sem paginação)" }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:  { type: array, items: { $ref: '#/components/schemas/EmpresaPendencias' } }
                  total:  { type: integer, description: nº de empresas que casam o filtro (antes do limit/offset) }
                  limit:  { type: integer }
                  offset: { type: integer }

  /nfse/import:
    get:
      summary: Status de IMPORTAÇÃO de NFSe (lado Firebird, sem etapa de chegada) por empresa/período
      description: NFSe não tem chave de 44 dígitos no filesystem (Fase 0) — só rastreamos o estado de import no Athenas.
      parameters:
        - { name: codigo_empresa, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [imported, import_ignored, pending_import] } }
        - { name: from,   in: query, schema: { type: string, format: date } }
        - { name: to,     in: query, schema: { type: string, format: date } }
        - { name: limit,  in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        athenas_chave: { type: string }
                        empresa:       { $ref: '#/components/schemas/EmpresaRef' }
                        status:        { type: string }
                        motivo_ignorado: { type: string, nullable: true }
                        data_emissao:  { type: string, format: date, nullable: true }
                        valor_total:   { type: number, nullable: true }
                  total: { type: integer }

  /metrics/overview:
    get:
      summary: Cards do dashboard (totais por etapa, em trânsito, importadas hoje, ignoradas, latências p50/p95)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  arrived:         { type: integer }
                  synced:          { type: integer }
                  imported_today:  { type: integer }
                  import_ignored:  { type: integer }
                  in_transit:      { type: integer }
                  stuck:           { type: integer }
                  lost:            { type: integer }
                  # Percentis sobre JANELA MÓVEL de 30 dias (arrived_at p/ chegada->sync,
                  # synced_at p/ sync->import). Exclui o backfill histórico, que inflava
                  # os números. Rótulo honesto na UI: "últimos 30 dias".
                  lat_arrival_sync_p50_s: { type: integer, nullable: true }
                  lat_arrival_sync_p95_s: { type: integer, nullable: true }
                  lat_sync_import_p50_s:  { type: integer, nullable: true }
                  lat_sync_import_p95_s:  { type: integer, nullable: true }

  /metrics/timeseries:
    get:
      summary: Série temporal da evolução do pipeline (Painel v2 — gráficos de linha de volume e latência por dia/semana)
      description: >
        Fluxo-por-evento: arrived/synced/imported contam notas cujo arrived_at/synced_at/imported_at
        caiu no bucket; import_ignored conta notas com status atual import_ignored, datadas pelo
        observed_at do evento de ignore (reconcilia com o card de /metrics/overview; exclui as
        corrigidas via --fix-pending). Latências são percentis (segundos) por coorte de evento
        (chegada->sync chaveada por quem chegou no bucket; sync->import por quem sincronizou) —
        os dias mais recentes podem ser parciais (censura à direita das coortes ainda em trânsito).
        Bucketização em America/Sao_Paulo. Série contínua: todo bucket do range aparece, com
        contagens zeradas e latências null quando não há dado.
      parameters:
        - { name: range,  in: query, schema: { type: string, enum: [7d, 30d, 90d], default: 30d } }
        - { name: bucket, in: query, schema: { type: string, enum: [day, week], default: day } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  range:  { type: string, example: 30d }
                  bucket: { type: string, example: day }
                  tz:     { type: string, example: America/Sao_Paulo }
                  buckets:
                    type: array
                    items:
                      type: object
                      properties:
                        date:           { type: string, format: date, description: "YYYY-MM-DD (fuso local); semana = segunda-feira" }
                        arrived:        { type: integer }
                        synced:         { type: integer }
                        imported:       { type: integer }
                        import_ignored: { type: integer }
                        lat_arrival_sync_p50_s: { type: integer, nullable: true }
                        lat_arrival_sync_p95_s: { type: integer, nullable: true }
                        lat_sync_import_p50_s:  { type: integer, nullable: true }
                        lat_sync_import_p95_s:  { type: integer, nullable: true }
        '400': { description: range ou bucket inválido }

  /metrics/doctypes:
    get:
      summary: Distribuição de notas por tipo de documento (NFe/NFCe/CTe/...)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        doc_type: { $ref: '#/components/schemas/DocType' }
                        count:    { type: integer }

  /metrics/backlog-age:
    get:
      summary: Notas pendentes por faixa de idade desde a chegada (notas presas na fila)
      description: >
        Faixas fixas (<1h, 1-6h, 6-24h, 1-3d, 3-7d, >7d), sempre todas presentes (0 quando vazia),
        sobre notas não-terminais (arrived/synced/pending_import). Idade = now − arrived_at
        (COALESCE com first_seen_at). Útil p/ ver notas presas aguardando sincronização.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        label: { type: string, example: "1-6h" }
                        count: { type: integer }

  /metrics/throughput:
    get:
      summary: "(placeholder antigo, não implementado — use /metrics/timeseries) Série temporal por etapa"
      parameters:
        - { name: bucket, in: query, schema: { type: string, enum: [hour, day], default: hour } }
        - { name: from,   in: query, schema: { type: string, format: date-time } }
        - { name: to,     in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      type: object
                      properties:
                        t:        { type: string, format: date-time }
                        arrived:  { type: integer }
                        synced:   { type: integer }
                        imported: { type: integer }

  /jobs/{maestroJobId}/notas:
    get:
      summary: Notas correlacionadas a um job do maestro (fase posterior — pode vir vazio no MVP)
      parameters:
        - { name: maestroJobId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: ok, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Nota' } }, total: { type: integer } } } } } }